backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
openai>=1.35.0
anthropic>=0.28.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.14
langchain-anthropic>=0.1.15
tiktoken>=0.7.0
PyGithub>=2.3.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
Pillow>=10.0.0
python-jose>=3.3.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
numpy>=1.26.0
python-pptx>=0.6.23
google-generativeai>=0.8.0
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.0.0
pocket-tts>=1.1.1
